+++ packages/web/src/App.vue
@@ -10,6 +10,7 @@ import {
   currentLanguage,
 } from './language';
 import Select from './components/Select.vue';
+import { switchChartTheme } from './views/chart/echart';
 const route = useRoute();
 
 // relationship packages hot word unknowns
@@ -20,24 +21,37 @@ const menus = [
   { name: $tf('隐式引用'), icon: 'icon-menu-unuse', path: '/unknowns' },
 ];
 
-const SIDEBAR_WIDTH = 120;
-
+const SIDEBAR_WIDTH = 140;
 const isActiveMenu = (path: string) => {
   return route.path === path;
 };
 
 const isDarkModel = ref(false);
-const onSwitchTheme = () => {
+
+const initTheme = () => {
+  const isDark = localStorage.getItem('theme') === 'dark';
+  onSwitchTheme(isDark);
+};
+
+const handleSwitchTheme = () => {
   const isDark = document.body.classList.contains('theme-dark');
-  if (isDark) {
+  onSwitchTheme(!isDark);
+};
+const onSwitchTheme = (dark: boolean) => {
+  switchChartTheme(dark ? 'dark' : 'default');
+  if (!dark) {
     isDarkModel.value = false;
     document.body.classList.remove('theme-dark');
+    localStorage.removeItem('theme');
   } else {
     isDarkModel.value = true;
     document.body.classList.add('theme-dark');
+    localStorage.setItem('theme', 'dark');
   }
 };
+
 onBeforeMount(() => {
+  initTheme();
   getConfig().then((res) => {
     window.CONFIG = res;
   });
@@ -68,7 +82,7 @@ function openProject() {
         :to="item.path"
       >
         <IconBtn :icon="item.icon" :active="isActiveMenu(item.path)"></IconBtn>
-        <span class="ml-1">{{ item.name }}</span>
+        <span class="ml-1 ui-text-justify flex-1">{{ item.name }}</span>
       </router-link>
       <div class="absolute left-0 bottom-0 w-full py-8 px-2">
         <Select
@@ -77,7 +91,7 @@ function openProject() {
           class="w-full"
           @onChange="(v) => switchLanguage(v)"
         ></Select>
-        <p class="my-4 flex">
+        <p class="my-4 flex justify-between">
           <IconBtn
             :icon="isDarkModel ? 'icon-settings-fill' : 'icon-settings-fill'"
             @click="openProject"
@@ -85,7 +99,7 @@ function openProject() {
           ></IconBtn>
           <IconBtn
             :icon="isDarkModel ? 'icon-dark' : 'icon-baitianmoshi'"
-            @click="onSwitchTheme"
+            @click="handleSwitchTheme"
           >
           </IconBtn>
         </p>
@@ -109,7 +123,7 @@ function openProject() {
 :root {
   --an-c-active: #ff7f50;
   --an-c-active-light: #e3d6d2;
-  --an-c-normal: #606266;
+  --an-c-normal: #434343;
   --an-c-black: #1a1a1a;
   --an-c-gray: #f0f2f7;
   --an-c-white: #ffffff;
@@ -208,5 +222,17 @@ button[disabled] {
   border: 1px solid var(--an-c-light);
   border-radius: 4px;
   transition: border-color 0.2s cubic-bezier(0.645, 0.045, 0.355, 1);
+}
+.ui-text-justify {
+  height: 30px;
+  line-height: 30px;
+  padding-left: 10px;
+  padding-right: 10px;
+  text-align: justify;
+}
+.ui-text-justify::after {
+  display: inline-block;
+  width: 100%;
+  content: '';
 }
 </style>