+++ packages/server/CHANGELOG.md
@@ -1,3 +1,64 @@
+# [2.8.0](https://github.com/chennlang/js-analyzer/compare/v2.7.0...v2.8.0) (2025-01-08)
+
+
+### Features
+
+* 切换打开编辑器的插件&支持在config中配置使用什么编辑器打开，默认值为vscode ([45dc22e](https://github.com/chennlang/js-analyzer/commit/45dc22e4ce6c95f399697ebe7f394130e5eec92e))
+
+# [2.7.0](https://github.com/chennlang/js-analyzer/compare/v2.6.2...v2.7.0) (2024-11-29)
+
+
+### Features
+
+* update node styles & add switch node display level ([3c6518d](https://github.com/chennlang/js-analyzer/commit/3c6518d49fd6ec308b9ffa430436a25b61b1bc58))
+
+## [2.6.2](https://github.com/chennlang/js-analyzer/compare/v2.6.1...v2.6.2) (2024-11-06)
+
+
+### Bug Fixes
+
+* 去除 core 自动发布，改为手动发布 ([320bf93](https://github.com/chennlang/js-analyzer/commit/320bf938bb17ba1f0738fd7c6ad997b1bb0daa00))
+
+## [2.5.1](https://github.com/chennlang/js-analyzer/compare/v2.5.0...v2.5.1) (2024-10-26)
+
+
+### Bug Fixes
+
+* 不支持 win & 局部配置支持相对路径 ([dea2ab7](https://github.com/chennlang/js-analyzer/commit/dea2ab72c57fa4d95957ebdac1e01c2e42d1782b))
+
+# [2.5.0](https://github.com/chennlang/js-analyzer/compare/v2.4.0...v2.5.0) (2024-08-27)
+
+
+### Features
+
+* application interface style optimization ([0aeb82e](https://github.com/chennlang/js-analyzer/commit/0aeb82e6ee757fce131fdc5d48bbff9c093a34e7))
+
+# [2.4.0](https://github.com/chennlang/js-analyzer/compare/v2.3.1...v2.4.0) (2024-07-25)
+
+
+### Bug Fixes
+
+* Cannot find module '../libs/core-dist/js-analyzer-core.cjs.js' ([2a1c47f](https://github.com/chennlang/js-analyzer/commit/2a1c47f0e3413a765b6fc48309f8c1e2bf509b1b))
+* Cannot find module js-analyzer-core.cjs.js ([9889be4](https://github.com/chennlang/js-analyzer/commit/9889be477fdce3e00d5c3cc47a8721e243d5e34e))
+* no such file or directory:public/data ([4dd5219](https://github.com/chennlang/js-analyzer/commit/4dd5219005301d362a9b41a2494e907246d94465))
+* not find public/data ([86c28ab](https://github.com/chennlang/js-analyzer/commit/86c28ab1bd878746ecfd165dd2da96df2f3a9165))
+* remove publish script ([77107da](https://github.com/chennlang/js-analyzer/commit/77107da09e72244badafdfc1d17a60b1daf303df))
+* remove test.json ([0107ae3](https://github.com/chennlang/js-analyzer/commit/0107ae3af7d9afb36ba8ff55bf396324d029e0c6))
+* web page js file 404 ([b04c1e3](https://github.com/chennlang/js-analyzer/commit/b04c1e309abd171f70125a5a27c210af73c46f18))
+
+
+### Features
+
+* 支持国际化语言 ([a778058](https://github.com/chennlang/js-analyzer/commit/a778058751f36d954ca9d16d055f446a2e43a684))
+* 添加日语文档 ([58b40e7](https://github.com/chennlang/js-analyzer/commit/58b40e78579ad4e101a5ec86676340b30c83ba2e))
+
+## [2.3.1](https://github.com/chennlang/js-analyzer/compare/v2.3.0...v2.3.1) (2024-07-05)
+
+
+### Bug Fixes
+
+* 修复本地启动异常 ([4c83592](https://github.com/chennlang/js-analyzer/commit/4c83592c78f91d033b69fa3665cbfc968e619be7))
+
 # [2.3.0](https://github.com/chennlang/js-analyzer/compare/v2.2.6...v2.3.0) (2024-06-26)